# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PORTAL API Documentation",
    "version": "1.0.0",
    "description": "Documentación técnica de la API de PORTAL - Sistema de Gestión de Personal y Horarios."
  },
  "servers": [{ "url": "http://localhost:4000", "description": "Servidor de Desarrollo" }],
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": { "success": { "type": "boolean" }, "message": { "type": "string" } },
        "required": ["success"]
      },
...
      "Employee": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "rut": { "type": "string", "minLength": 8 },
          "name": { "type": "string", "minLength": 2 },
          "position": { "type": "string" },
          "department": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "emergencyContact": { "type": "string" },
          "status": { "type": "string", "enum": ["Activo", "Archivado"] },
          "contractType": { "type": "string" },
          "shiftPatternId": { "type": "string" },
          "hireDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "terminationDate": { "type": "string", "nullable": true },
          "address": { "type": "string" },
          "commune": { "type": "string" },
          "region": { "type": "string" },
          "birthDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "gender": { "type": "string" },
          "bankName": { "type": "string" },
          "accountType": { "type": "string" },
          "accountNumber": { "type": "string" },
          "createUserAccount": { "type": "boolean" },
          "isDeleted": { "type": "boolean" },
          "syncStatus": { "type": "string" },
          "lastModified": { "type": "number" }
        },
        "required": ["rut", "name"]
      },
...
      "KioskLogin": {
        "type": "object",
        "properties": { "employeeId": { "type": "string" }, "pin": { "type": "string" } },
        "required": ["employeeId", "pin"]
      },
      "Login": {
        "type": "object",
        "properties": { "username": { "type": "string" }, "password": { "type": "string" } },
        "required": ["username", "password"]
      },
...
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "timestamp": { "type": "string" },
          "actorUsername": { "type": "string" },
          "action": { "type": "string" },
          "category": { "type": "string" },
          "severity": { "type": "string" },
          "outcome": { "type": "string" },
          "details": { "type": "object", "additionalProperties": { "nullable": true } },
          "ipAddress": { "type": "string", "nullable": true },
          "metadata": { "type": "object", "additionalProperties": { "nullable": true } }
        },
        "required": [
          "id",
          "timestamp",
          "actorUsername",
          "action",
          "category",
          "severity",
          "outcome"
        ]
      },
      "ConfigEntry": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "value": { "nullable": true },
          "group": { "type": "string" },
          "updatedAt": { "type": "string" }
        },
        "required": ["key"]
      },
...
      "ServerTime": {
        "type": "object",
        "properties": {
          "iso": { "type": "string" },
          "timestamp": { "type": "number" },
          "timezone": { "type": "string" },
          "businessDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" }
        },
        "required": ["iso", "timestamp", "timezone", "businessDate"]
      },
...
      "AuditLogListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiResponse" },
          {
            "type": "object",
            "properties": {
              "pagination": {
                "type": "object",
                "properties": {
                  "total": { "type": "number" },
                  "page": { "type": "number" },
                  "totalPages": { "type": "number" }
                },
                "required": ["total", "page", "totalPages"]
              },
              "data": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLog" } }
            },
            "required": ["pagination", "data"]
          }
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/configs/company-policy": {
      "post": {
        "summary": "Subcribir la politica corporativa en PDF",
        "tags": ["Configs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Politica corporativa actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/configs/server-time": {
      "get": {
        "summary": "Obtener hora actual del servidor",
        "tags": ["Config"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Hora del servidor",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ServerTime" } }
            }
          }
        }
      }
    },
    "/api/configs/validate-closure": {
      "get": {
        "summary": "Validar si una fecha puede cerrar el periodo contable",
        "tags": ["Configs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "date", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "Resultado de la validacion de cierre",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/configs": {
      "get": {
        "summary": "Listar configuraciones del sistema",
        "tags": ["Configs"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de configuraciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ConfigEntry" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/configs/{key}": {
      "get": {
        "summary": "Obtener una configuración por llave",
        "tags": ["Config"],
        "parameters": [
          { "in": "path", "name": "key", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Entrada de configuración",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ConfigEntry" } }
            }
          }
        }
      },
      "post": {
        "summary": "Establecer una configuración por llave",
        "tags": ["Config"],
        "parameters": [
          { "in": "path", "name": "key", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "value": { "type": "any" } } }
            }
          }
        },
        "responses": { "200": { "description": "Configuración guardada" } }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Inicia sesión de usuario (Soporta MFA)",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Login" } } }
        },
        "responses": {
          "200": {
            "description": "Login exitoso o requerimiento de MFA",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "token": { "type": "string" },
                    "requiresMFA": { "type": "boolean" },
                    "mustChangePassword": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "Credenciales inválidas" }
        }
      }
    },
    "/api/auth/mfa/setup": {
      "post": {
        "summary": "Genera secreto MFA y QR",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Secreto generado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "qrCode": { "type": "string" }, "secret": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/mfa/verify": {
      "post": {
        "summary": "Habilita MFA tras verificar primer código",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "token": { "type": "string" } } }
            }
          }
        },
        "responses": { "200": { "description": "MFA habilitado" } }
      }
    },
    "/api/auth/mfa/validate": {
      "post": {
        "summary": "Valida código MFA durante login",
        "tags": ["Auth"],
        "responses": {
          "200": {
            "description": "Autenticación de dos pasos exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "mustChangePassword": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "Código o desafío inválido, expirado o MFA deshabilitado" },
          "403": { "description": "Cuenta archivada entre factores" },
          "429": {
            "description": "Cinco códigos incorrectos bloquean MFA durante cinco minutos",
            "headers": {
              "Retry-After": {
                "description": "Segundos hasta que se permita otro intento",
                "schema": { "type": "integer", "minimum": 1 }
              }
            }
          }
        }
      }
    },
    "/api/auth/kiosk-login": {
      "post": {
        "summary": "Inicia sesión modo Kiosko (Reloj Control)",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/KioskLogin" } }
          }
        },
        "responses": {
          "200": {
            "description": "Login de Kiosko exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "employee": { "$ref": "#/components/schemas/Employee" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audit-logs/export": {
      "get": {
        "summary": "Exportar logs de auditoría",
        "tags": ["AuditLogs"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Archivo de auditoría generado" } }
      }
    },
    "/api/audit-logs": {
      "get": {
        "summary": "Obtener logs de auditoría paginados",
        "tags": ["AuditLogs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "string" } },
          { "in": "query", "name": "pageSize", "schema": { "type": "string" } },
          { "in": "query", "name": "category", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de logs de auditoría",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuditLogListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear log de auditoría manual",
        "tags": ["AuditLogs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AuditLog" } }
          }
        },
        "responses": { "201": { "description": "Log creado" } }
      }
    },
    "/api/audit-logs/integrity-status": {
      "get": {
        "summary": "Obtener estado de integridad de la auditoría",
        "tags": ["AuditLogs"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Reporte de integridad" } }
      }
    },
    "/api/audit-logs/verify-integrity": {
      "get": {
        "summary": "Verificar integridad de la auditoría (Elevado)",
        "tags": ["AuditLogs"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Verificación iniciada" } }
      }
    },
    "/api/audit-logs/cleanup": {
      "post": {
        "summary": "Limpiar logs de auditoría antiguos",
        "tags": ["AuditLogs"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Limpieza realizada" } }
      }
    },
...
  },
...
}